Update Seeders, Entities, DTOs. Create some Controllers, Entity Mapping . .

+++ LanGeng.API/Seeders/DatabaseSeeder.cs
@@ -8,8 +8,11 @@ public static class DatabaseSeeder
     public static async Task Seed(IServiceProvider serviceProvider)
     {
         using var context = new SocialMediaDatabaseContext(serviceProvider.GetRequiredService<DbContextOptions<SocialMediaDatabaseContext>>());
-        context.Database.EnsureDeleted();
-        await context.Database.MigrateAsync();
+        if (context.Database.GetPendingMigrations().Count() > 0)
+        {
+            context.Database.EnsureDeleted();
+            await context.Database.MigrateAsync();
+        }
         await UserSeeder.Seed(context);
         await UserPostSeeder.Seed(context);
     }
